+++ src/test.js
@@ -4,6 +4,7 @@ require("./data/url/data-url_test");
 require("./data/parse/parse_test");
 
 require("./data/localstorage-cache/localstorage-cache_test");
+require("./data/memory-cache/memory-cache_test");
 require("./data/combine-requests/combine-requests_test");
 require("./data/inline-cache/inline-cache_test");
